Move classes added in #511 to implcore package.
