+++ core_impl/src/test/java/io/opencensus/implcore/tags/TagContextDeserializationTest.java
@@ -11,12 +11,12 @@
  * limitations under the License.
  */
 
-package io.opencensus.impl.tags;
+package io.opencensus.implcore.tags;
 
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.common.collect.Lists;
-import io.opencensus.impl.internal.VarInt;
+import io.opencensus.implcore.internal.VarInt;
 import io.opencensus.tags.TagContext;
 import io.opencensus.tags.TagContextBinarySerializer;
 import io.opencensus.tags.TagContexts;
